Kernel: Lock region tree while accessing it in page fault handler
Before this change, we were happily grabbing a raw pointer to the
faulting region from the region tree, and then using it without
protection. This was racy, as another thread could call munmap() on
the same region while we were processing a page fault inside it.

After this patch, page fault handling now works like this:

1. Lock the region tree
2. Find the faulting region & VM object, perform access checks
3. Unlock the region tree
4. Abort with SIGSEGV if no region found, or access checks failed
5. Have VM object load/allocate/cow the faulting page as appropriate
6. Relock the region tree
7. Find the fault region & VM object
8. If the found VM object is the same as before, update page tables
9. Unlock the region tree

In other words, we unlock the region tree while letting the VM object
do the appropriate work of "faulting in" the physical page.
This is necessary because some VM objects (e.g inode VM objects) may
need to block on I/O.