LibWeb: Make `resolution` media queries actually work
For these to behave as expected, we have to communicate the current
screen DPI from the UI process to the WC process. This is done with
a new set_screen_dpi(double) IPC API.

If no value is provided (or a media query executes in a detached
context somehow) we fall back to 96 dpi as a default value.

This makes high resolution images appear on https://apple.com/ :^)